% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embedding.R
\name{runEmbedding}
\alias{runEmbedding}
\title{runEmbedding}
\usage{
runEmbedding(req, data)
}
\arguments{
\item{req}{type = (tsne,umap,pca)
umap config = (minimumDistance, distanceMetric)
tsne config = (perplexity,learningRate)}

\item{data}{Object of class Seurat}
}
\value{
x,y positions for each cell, sorted by cellId with NA values in filtered cells.
}
\description{
runEmbedding
}
